+++ test.csv
@@ -1,0 +1,3 @@
+username,password,service
+apina,123412345a,foobar.net
+apina,123412345a,foobar2.net
